Added a group of new "Valley" plots.
These plot show instances from the 10K run where the fitness got worse
before getting (substantially) better. There are in fact the 7 sections
of the root parent path that went up before going down and have the
best per-generation improvement relative to the fitness at the start of
the path.